For DHW recirculation measure case, added flow schedules for mass flow rate, temperature and ambient air velocity
For SWWH015 the mass flow rate schedule is used to achieve
the measure case control goals.
The mass flow schedule is applied to recirculation (LoadProfile) pump in the
LoadProfile branch.

For SWWHC016 the temperature schedule is used to achieve the measure case
control goals.

Ambient air velocity is used to model realistic heat transfer for pipe losses.